Remote volume control
Special case situation: we have pico media remotes that control Sonos devices, but the volume control is for that room and not the entire house.  This optional feature looks for a volume change anywhere in the group and applies that delta to each device the group.  To Do: validate/debug, add disable switching in HomeKit